+++ src/compas_view2/shapes/text.py
@@ -7,9 +7,10 @@ class Text(Shape):
     A text is defined by a string of characters, a position, a height, and a font.
     """
 
-    def __init__(self, text, position=[0, 0, 0], height=50, font=None):
+
+    def __init__(self, text, position=[0, 0, 0], height=50, absolute_height=False):
         super().__init__()
         self.text = text
         self.position = Vector(*position)
         self.height = height
-        self.font = font
+        self.absolute_height = absolute_height