+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Text font option.
-* Multi-cursor effect. 
+- Text absolute height option.
+
 ### Changed
 
 ### Removed